feat: 대화 초기화 기능 추가 및 복사 버튼 컴포넌트 분리 (#171)
